Moving config logic to separate module
