+++ config/src/test/groovy/com/okta/spring/config/DiscoveryMetadataTest.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.okta.spring.oauth.discovery
+package com.okta.spring.config
 
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpInputMessage
