+++ config/src/test/resources/discovery-test.json
