+++ pom.xml
@@ -36,6 +36,7 @@
     </properties>
 
     <modules>
+        <module>config</module>
         <module>oauth2</module>
         <module>okta-spring-boot-starter</module>
         <module>examples</module>
@@ -87,6 +88,11 @@
             <dependency>
                 <groupId>com.okta.spring</groupId>
                 <artifactId>okta-spring-security-oauth2</artifactId>
+                <version>0.2.0-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>com.okta.spring</groupId>
+                <artifactId>okta-spring-config</artifactId>
                 <version>0.2.0-SNAPSHOT</version>
             </dependency>
